Simplify the bcbio cluster bootstrap
* Remove bcbiovm.provider.aws.bootstrap module
* Remove bcbiovm.provider.azure.bootstrap module
* Add bcbiovm.provider.common module
* Remove duplicate code blocks
