+++ bcbiovm/provider/common/playbook.py
@@ -11,6 +11,7 @@ class Playbook(object):
     bcbio = ("roles", "bcbio_bootstrap", "tasks", "main.yml")
     docker = ("roles", "docker", "tasks", "main.yml")
     docker_local = ("bcbio_vm_docker_local.yml", )
+    gof3r = ("roles", "gof3r", "tasks", "main.yml")
     nfs = ("roles", "encrypted_nfs", "tasks", "main.yml")
 
     def __getattribute__(self, name):
@@ -24,18 +25,3 @@ class Playbook(object):
             setattr(self, name, attribute)
 
         return attribute
-
-
-class AWSPlaybook(Playbook):
-
-    """Default paths for Ansible playbooks."""
-
-    gof3r = ("roles", "gof3r", "tasks", "main.yml")
-    icel = ("roles", "icel", "tasks", "main.yml")
-    mount_lustre = ("roles", "lustre_client", "tasks", "mount.yml")
-    unmount_lustre = ("roles", "lustre_client", "tasks", "unmount.yml")
-
-
-class AzurePlaybook(Playbook):
-
-    """Default paths for Ansible playbooks."""
